Use a monotonically increasing timer
gettimeofday() can flip around all willy-nilly, and is actually common for
users to cause this by having memcached start on boot before ntp launches
and corrects system time. libevent fires events on a monotonically increasing
clock, so we can more or less safely tick our internal timer up by one second
every time the handler runs.

Unfortunately we support expiration by date. If memcached's start time isn't
in sync with the rest of the world this feature won't work well, but it never
did.

Was originally going to make this optional, but I can't come up with a great
reason to do so. If it turns out this isn't "accurate enough", we can add the
clock_gettime() code inline.