slab rebalancing from random class
specifying -1 as the src class for a slabs reassign will pick the first
available, rolling through the list on each request (so as to not bias against
the smaller classes).

So if you're in a hurry and have to move memory into class 5, you may now mash
it without thinking.