slab_reassign can perform a shrink
accpets an additional parameter to indicate several slabs which go the same way
(currently ignored for reassignment).
dest<0 in the runtime is used to indicate the number of slabs, otherwise it is 1.

The rebalance process (start and finish) also support a num_slabs of shrinkage operations.
This might happen faster if instead of releasing one by one they were all done together.