automove_decision_maker and the activating thread support shrinkage
Decision is taken on the basis of automove_level (until level 2),
where 2 is agressive and 1 is easy-going.
When it is agressive, it uses an emergency source - the best recomendation,
meeting only the criterion of enough slabs.
Also, multi-slab at one go is supported, with a conservative algorithm
for estimating the minimal amount to take from the first one, and then the
situation is re-estimated.