bugfixed and improvements in shrinkage:
Inform the user about the real number of slabs which need to be released.
Count the slablist , hashtable and slabs memory, and combine them in TOTAL_MALLOCED.
defining DEBUG_SLABS activates intense printing reports on memory status.
slab list is only grown if threre is enough room to grow it, and then the memory is checked again - it could be that there is only room for either growing the slab list or adding another slab.
When the eviction statistics are not helpful, the number of slabs in the class is used as a tie breaker.
bugfix: only a positive mem gap should lead to intense shrink, not just any non-zero gap.
bugfix: only slabs with more than one slab are now counted when the number of classes is used to compute the average shrink.